refactor: 디렉토링 수정 및 네이밍 수정
